fix(sms): Escape newline characters when sending text messages
CFML does not automatically escape newline characters in strings.
This means when encoding a string with `\n` in it, it encodes the
backslash as its own character.  This means instead of getting the
encoded value of `%0A` (newline) we get `%5Cn` which is `\n`. We
handle the escaping of newline characters when sending through the
`sms` endpoint.  Other locations should either 1) send a PR to this
library to handle it correctly or 2) use `chr( 10 )` inside their
string instead of `\n`.